requirements.txt
flask
flask[async]
python-dotenv
openai>=1.0.0
anthropic>=0.3.11
numpy
agno
streamlit
playwright
bs4